pyproject.toml
[project]
name = "kobeck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Lukas Anzinger", email = "lukas@lukasanzinger.at" }
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]

[project.scripts]
kobeck = "kobeck:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fastapi-cli>=0.0.10",
    "ruff>=0.12.12",
]
